split tests consistently with modules
